Rename "util" package to "composables"
Rename the "util" package to "composables" and move all non-composable
helpers to a new "utils" package and the "model" package.
