+++ src/main/kotlin/composables/CaptionedText.kt
@@ -1,4 +1,4 @@
-package org.ossreviewtoolkit.workbench.util
+package org.ossreviewtoolkit.workbench.composables
 
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
