+++ src/main/kotlin/composables/FileDialog.kt
@@ -1,4 +1,4 @@
-package org.ossreviewtoolkit.workbench.util
+package org.ossreviewtoolkit.workbench.composables
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.awt.ComposeWindow
