+++ src/main/kotlin/ui/violations/ViolationsViewModel.kt
@@ -12,13 +12,13 @@ import org.ossreviewtoolkit.model.Severity
 import org.ossreviewtoolkit.utils.spdx.SpdxSingleLicenseExpression
 import org.ossreviewtoolkit.workbench.model.FilterData
 import org.ossreviewtoolkit.workbench.model.OrtModel
+import org.ossreviewtoolkit.workbench.model.ResolutionStatus
 import org.ossreviewtoolkit.workbench.model.Violation
-import org.ossreviewtoolkit.workbench.util.ResolutionStatus
-import org.ossreviewtoolkit.workbench.util.SpdxExpressionStringComparator
-import org.ossreviewtoolkit.workbench.util.matchResolutionStatus
-import org.ossreviewtoolkit.workbench.util.matchString
-import org.ossreviewtoolkit.workbench.util.matchStringContains
-import org.ossreviewtoolkit.workbench.util.matchValue
+import org.ossreviewtoolkit.workbench.utils.SpdxExpressionStringComparator
+import org.ossreviewtoolkit.workbench.utils.matchResolutionStatus
+import org.ossreviewtoolkit.workbench.utils.matchString
+import org.ossreviewtoolkit.workbench.utils.matchStringContains
+import org.ossreviewtoolkit.workbench.utils.matchValue
 
 class ViolationsViewModel(private val ortModel: OrtModel = OrtModel.INSTANCE) {
     private val scope = CoroutineScope(Dispatchers.Default)
